Refactor use of `default` template function
We have a lot of uses of the `default` template function like:

  {{ default A B }}

which is a little non-obvious that A is the default value. Replace these
with:

  {{ B | default A }}

------------------------------------------------------------

We also have some instances of:

  {{ default A B C }}

which is invalid. Replace those with:

  {{ coalesce A B C }}

-------------------------------------------------------------

We also have some instances of:

  {{ coalesce A B }}

For consistency's sake, replace those with:

  {{ A | default B }}

JIRA: INFRA-28410