Allow using host network on main deployment
A limitation with the AWS VPC CNI is that we have to put a webhook
pod on the host's network to allow the EKS control plane to communicate
with it.

This commit allows the pods of the main deployment to run on the host
network so we can deploy webhooks with the standard-application-stack
chart.