Allow extraIngresses to override any field of the top-level ingress
I had a problem where settings `ingress.extraIngresses.pathBasedRouting`
was ignored because the extra ingresses would always use
`ingress.pathBasedRouting`. This commit fixes that by allowing the extra
ingresses to generically override any field of the top-level ingress
i.e. the top-level ingress values are used as the defaults, which can be
overridden.